# Make deep expressions traverse OtherStmt, Lambda nodes