--- a/semgrep-core/tests/go/deep_expr_lambda.go
+++ b/semgrep-core/tests/go/deep_expr_lambda.go
@@ -1,0 +1,4 @@
+// ERROR: match
+x := 1
+go func() { fmt.Println(x) } ()
+