… semgrep-core/tests/go/deep_expr_lambda.sgrep