A few easily recurseable cases were omitted from deep-statement
matching: OtherStmt nodes, and Lambda nodes. The former can refer to an
arbitrary number of `any` nodes, and we update to traverse any wrapped
`E` nodes. The latter we treat like any other `FunctionDef` node.

The `OtherStmt` change should be relatively innocuous, as these
statements are relatively rare in code. I'm a bit worried about the
performance impacts of the Lambda change, however, as a minified js
file could easily contain a very large number of anonymous functions.

To that end, let's not merge this until we can confirm that we are not
creating a performance regression with this change.

Fixes #2879.

Fixes #2878.

Perhaps not the most complete argument re. performance, but looking at `p/nodejsscan` (which makes heavy use of deep expressions) against `angular.min.js` (which makes heavy use of anonymous functions) yields no significant regression:

Before this change:
```
semgrep --config p/nodejsscan node_modules/angular/angular.min.js  233.27s user 12.69s system 98% cpu 4:08.47 total
```

After this change:
```
semgrep --config p/nodejsscan node_modules/angular/angular.min.js  229.75s user 12.45s system 98% cpu 4:05.77 total
```


PR checklist:
- [x] changelog is up to date